Enhance error analysis capabilities in metrics module
This commit introduces two new functions: `plot_error_analysis_chart` and `generate_error_analysis_report`. The `plot_error_analysis_chart` function visualizes prediction errors relative to predicted probabilities, supporting both binary and multi-class classification. The `generate_error_analysis_report` function provides a tabular report that groups predictions by feature values and computes error metrics per group.

Key updates:
- Documentation enhancements in README.md, skills/metrics/SKILL.md, and relevant RST files to include usage examples and detailed descriptions.
- Version bump to 1.10.0rc9 to reflect the addition of new features.
- Removal of outdated test images and tests related to the previous error analysis implementations.

These changes improve the usability and functionality of the error analysis tools within the library.
