+++ ds_utils/metrics/__init__.py
@@ -3,6 +3,7 @@
 This package provides functions for evaluating and visualizing model performance:
 - confusion_matrix - Confusion matrix visualization and classification metrics
 - curves - ROC and Precision-Recall curves with threshold annotations
+- error_analysis - Tabular error-analysis report generation
 - learning_curves - Learning curve visualization
 - probability_analysis - Probability calibration and accuracy analysis
 """
