+++ ds_utils/xai.py
@@ -7,9 +7,7 @@ import warnings
 
 from matplotlib import axes, image, pyplot as plt
 import numpy as np
-import pandas as pd
 import pydotplus
-import seaborn as sns
 from sklearn.tree import BaseDecisionTree, _tree as sklearn_tree, export_graphviz
 
 
@@ -184,190 +182,3 @@ def plot_features_importance(
     plt.xticks(rotation=90)
 
     return ax
-
-
-def plot_error_analysis_chart(
-    y_true: Union[np.ndarray, List],
-    y_pred: Union[np.ndarray, List],
-    y_proba: Union[np.ndarray, List],
-    positive_class: Union[int, str],
-    *,
-    classes: Optional[List] = None,
-    ax: Optional[axes.Axes] = None,
-    **kwargs,
-) -> axes.Axes:
-    """Plot an error analysis chart showing prediction errors relative to predicted probabilities.
-
-    Builds an internal DataFrame with the predicted probability for the ``positive_class``
-    and an ``error_type`` column (correct, false_positive, false_negative), and draws
-    a violin plot showing the distribution of predicted probabilities across these error types.
-
-    For **binary** classification ``y_proba`` should be 1-D (probability of the positive class).
-    For **multi-class** classification ``y_proba`` should be 2-D with shape ``(n_samples, n_classes)``;
-    the column corresponding to ``positive_class`` is determined via ``classes``
-    (or inferred from ``np.unique(y_true)`` when ``classes`` is ``None``).
-    Error types are computed using a one-vs-rest scheme against ``positive_class``
-    by comparing ``y_true`` and ``y_pred`` directly. The original spec included a
-    ``threshold`` float parameter for re-thresholding raw probabilities; this
-    implementation intentionally omits it — callers should apply any desired
-    threshold to produce ``y_pred`` before calling this function.
-
-    :param y_true: Array-like of true labels.
-    :param y_pred: Array-like of predicted labels.
-    :param y_proba: Array-like of predicted probabilities.  1-D for binary or
-                    2-D with shape ``(n_samples, n_classes)`` for multi-class.
-    :param positive_class: The class to treat as the positive class.
-    :param classes: Ordered list of class labels matching the columns of ``y_proba``
-                    when it is 2-D.  If ``None``, inferred from ``np.unique(y_true)``.
-    :param ax: Axes object to draw the plot onto, otherwise uses the current Axes.
-    :param kwargs: Additional keyword arguments passed to :func:`seaborn.violinplot`.
-    :return: Axes object with the plot drawn onto it.
-    :raises ValueError: If inputs have mismatched lengths, ``y_proba`` has an invalid
-                        number of dimensions, or ``positive_class`` is not found in ``classes``.
-    """
-    y_true = np.asarray(y_true)
-    y_pred = np.asarray(y_pred)
-    y_proba = np.asarray(y_proba)
-
-    if y_true.shape[0] != y_pred.shape[0]:
-        raise ValueError("y_true and y_pred must have the same number of samples")
-    if y_true.shape[0] != y_proba.shape[0]:
-        raise ValueError("y_true and y_proba must have the same number of samples")
-
-    # Determine probability column for the positive class
-    if y_proba.ndim == 1:
-        proba = y_proba
-    elif y_proba.ndim == 2:
-        if classes is None:
-            classes = list(np.unique(y_true))
-        if positive_class not in classes:
-            raise ValueError(f"positive_class {positive_class!r} not found in classes {classes}")
-        class_index = classes.index(positive_class)
-        proba = y_proba[:, class_index]
-    else:
-        raise ValueError("y_proba must be 1-D or 2-D")
-
-    # Compute error_type using one-vs-rest logic on positive_class
-    is_positive_true = y_true == positive_class
-    is_positive_pred = y_pred == positive_class
-
-    error_type = np.select(
-        [
-            is_positive_true & ~is_positive_pred,
-            ~is_positive_true & is_positive_pred,
-        ],
-        ["false_negative", "false_positive"],
-        default="correct",
-    )
-
-    error_df = pd.DataFrame({"y_proba": proba, "error_type": error_type})
-
-    if ax is None:
-        _, ax = plt.subplots()
-
-    # order is hardcoded so the axis layout is consistent even when one
-    # error category is absent (e.g. a perfect model produces no FP/FN).
-    sns.violinplot(
-        x="error_type",
-        y="y_proba",
-        data=error_df,
-        order=["correct", "false_positive", "false_negative"],
-        ax=ax,
-        **kwargs,
-    )
-    if not ax.get_title():
-        ax.set_title(f"Error Analysis — positive class: {positive_class!r} (one-vs-rest)")
-    ax.set_xlabel("Error Type")
-    ax.set_ylabel("Predicted Probability")
-    ax.grid(axis="y", linestyle="--", alpha=0.7)
-
-    return ax
-
-
-def generate_error_analysis_report(
-    X: pd.DataFrame,
-    y_true: np.ndarray,
-    y_pred: np.ndarray,
-    feature_columns: Optional[List[str]] = None,
-    bins: int = 10,
-    threshold: float = 0.5,
-    min_count: int = 1,
-    sort_metric: str = "error_rate",
-    ascending: bool = False,
-) -> pd.DataFrame:
-    """Generate a tabular error-analysis report grouped by feature values.
-
-    The report groups predictions by feature values and computes error metrics per group.
-    For numerical features, values are binned into equal-width bins. For categorical
-    features, raw values are used as groups.
-
-    :param X: Feature DataFrame.
-    :param y_true: True labels.
-    :param y_pred: Predicted labels.
-    :param feature_columns: List of columns to analyze. If None, all columns in X are used.
-    :param bins: Number of bins for numerical features.
-    :param threshold: Threshold for probability-based error definitions.
-                      Validated but not used in the current implementation;
-                      reserved for future probability-based error definitions.
-    :param min_count: Minimum number of samples in a group to be included in the report.
-    :param sort_metric: Metric to sort the report by. Valid options are:
-                        'feature', 'group', 'count', 'error_count', 'error_rate', 'accuracy'.
-    :param ascending: Whether to sort in ascending order.
-    :return: DataFrame containing the error analysis report.
-    :raises ValueError: If bins < 1, threshold not in [0, 1], min_count < 1, or invalid sort_metric.
-    :raises KeyError: If any column in feature_columns is missing from X.
-    """
-    if bins < 1:
-        raise ValueError("bins must be at least 1")
-    if not (0 <= threshold <= 1):
-        raise ValueError("threshold must be between 0 and 1 inclusive")
-    if min_count < 1:
-        raise ValueError("min_count must be at least 1")
-
-    valid_columns = ["feature", "group", "count", "error_count", "error_rate", "accuracy"]
-    if sort_metric not in valid_columns:
-        raise ValueError(f"sort_metric must be one of {valid_columns}")
-
-    y_true = np.asarray(y_true)
-    y_pred = np.asarray(y_pred)
-
-    if len(y_true) != len(X) or len(y_pred) != len(X):
-        raise ValueError("X, y_true, and y_pred must have the same number of samples")
-
-    if feature_columns is not None:
-        missing_cols = [col for col in feature_columns if col not in X.columns]
-        if missing_cols:
-            raise KeyError(f"The following columns are missing from X: {missing_cols}")
-        cols_to_use = feature_columns
-    else:
-        cols_to_use = X.columns.tolist()
-
-    internal_df = X[cols_to_use].copy()
-    internal_df["__is_error__"] = y_true != y_pred
-
-    all_reports = []
-    for col in cols_to_use:
-        col_series = internal_df[col]
-        if pd.api.types.is_numeric_dtype(col_series):
-            groups = pd.cut(col_series, bins=bins)
-        else:
-            groups = col_series
-
-        report = (
-            internal_df.groupby(groups, observed=True)
-            .agg(count=("__is_error__", "size"), error_count=("__is_error__", "sum"))
-            .reset_index()
-        )
-        report.rename(columns={col: "group"}, inplace=True)
-        report["feature"] = col
-        report["error_rate"] = report["error_count"] / report["count"]
-        report["accuracy"] = 1 - report["error_rate"]
-
-        report = report[report["count"] >= min_count]
-        all_reports.append(report[valid_columns])
-
-    if not all_reports:
-        return pd.DataFrame(columns=valid_columns)
-
-    final_report = pd.concat(all_reports).sort_values(by=sort_metric, ascending=ascending).reset_index(drop=True)
-    return final_report
