+++ skills/metrics/SKILL.md
@@ -1,7 +1,7 @@
 ---
 name: ds-utils-metrics
 description: >
-  Provides evaluation metrics and visualization charts for machine learning models. Use when the user asks to evaluate a model, wants to plot a confusion matrix, ROC curve, or Precision-Recall curve, or needs to analyze learning curves or probability distributions in a Python data science project using sklearn-compatible models.
+  Provides evaluation metrics and visualization charts for machine learning models. Use when the user asks to evaluate a model, wants to plot a confusion matrix, ROC curve, or Precision-Recall curve, needs to analyze learning curves, probability distributions, or error analysis in a Python data science project using sklearn-compatible models.
 license: MIT
 metadata:
   author: Idan Morad
@@ -29,6 +29,8 @@ from ds_utils.metrics.learning_curves import plot_metric_growth_per_labeled_inst
 from ds_utils.metrics.probability_analysis import visualize_accuracy_grouped_by_probability
 from ds_utils.metrics.curves import plot_roc_curve_with_thresholds_annotations
 from ds_utils.metrics.curves import plot_precision_recall_curve_with_thresholds_annotations
+from ds_utils.metrics.probability_analysis import plot_error_analysis_chart
+from ds_utils.metrics.error_analysis import generate_error_analysis_report
 ```
 
 ---
@@ -216,6 +218,123 @@ fig.show()
 
 ---
 
+## plot_error_analysis_chart
+
+Automates the creation of an error analysis DataFrame (computing correct, false_positive, false_negative) and visualizes prediction errors relative to predicted probabilities. Supports both binary and multi-class classification using a one-vs-rest scheme.
+
+### Binary classification example
+
+```python
+from ds_utils.metrics.probability_analysis import plot_error_analysis_chart
+import matplotlib.pyplot as plt
+
+# complete usage example (binary)
+y_pred = clf.predict(X_test)
+y_proba = clf.predict_proba(X_test)[:, 1]  # probability of the positive class
+
+plot_error_analysis_chart(y_test, y_pred, y_proba, positive_class=1)
+plt.show()
+```
+
+### Multi-class classification example
+
+```python
+from ds_utils.metrics.probability_analysis import plot_error_analysis_chart
+import matplotlib.pyplot as plt
+
+# complete usage example (multi-class)
+y_pred = clf.predict(X_test)
+y_proba = clf.predict_proba(X_test)
+
+plot_error_analysis_chart(
+    y_test, y_pred, y_proba,
+    positive_class=1,
+    classes=clf.classes_.tolist()
+)
+plt.show()
+```
+
+**Parameters:**
+- `y_true` — array-like, True labels.
+- `y_pred` — array-like, Predicted labels (required).
+- `y_proba` — array-like, Predicted probabilities. 1-D for binary, 2-D `(n_samples, n_classes)` for multi-class.
+- `positive_class` — The class to treat as positive (used for correct/false_positive/false_negative assignment).
+- `classes` — list, optional. Ordered class labels matching columns of `y_proba` when 2-D. If `None`, inferred from `np.unique(y_true)`.
+- `ax` — matplotlib Axes, optional. Target axes for rendering.
+- `**kwargs` — forwarded to `seaborn.violinplot`.
+
+**Returns:** matplotlib Axes.
+
+**Common mistakes:**
+- Forgetting to pass `classes` for multi-class when the class order in `y_proba` does not match `np.unique(y_true)`. Always pass `classes=clf.classes_.tolist()` to be safe.
+- For binary classification, pass only the positive class probability column (1-D), not the full 2-D probability matrix, unless you also specify `classes`.
+- `y_pred` is required — you must pass pre-computed predictions, not raw probabilities.
+
+---
+
+## generate_error_analysis_report
+
+Provides a tabular error-analysis report that groups predictions by feature values and computes error metrics per group.
+
+```python
+import pandas as pd
+import numpy as np
+from ds_utils.metrics.error_analysis import generate_error_analysis_report
+
+# Setup dummy data with numerical and categorical features
+X_test = pd.DataFrame({
+    "age": [25, 30, 45, 50, 22, 35, 40, 60],
+    "region": ["North", "South", "North", "West", "East", "South", "West", "North"]
+})
+y_test = np.array([0, 1, 0, 1, 0, 1, 0, 1])
+y_pred = np.array([0, 1, 1, 1, 0, 0, 0, 1]) # Errors at index 2 and 5
+
+# complete usage example
+report_df = generate_error_analysis_report(
+    X_test, y_test, y_pred,
+    feature_columns=["age", "region"],
+    bins=3,
+    min_count=1,
+    sort_metric="error_rate",
+    ascending=False
+)
+print(report_df.head())
+```
+
+**Parameters:**
+- `X` — pandas DataFrame, Feature values.
+- `y_true` — array-like, True labels.
+- `y_pred` — array-like, Predicted labels.
+- `feature_columns` — list, optional. Subset of columns to analyze. If `None`, all columns in `X` are used.
+- `bins` — int, default 10. Number of bins for numerical features.
+- `threshold` — float, default 0.5. Reserved for future probability-based error definitions.
+- `min_count` — int, default 1. Minimum samples per group to include in the report.
+- `sort_metric` — str, default "error_rate". Column to sort by.
+- `ascending` — bool, default False. Sort direction.
+
+**Returns:** pandas DataFrame.
+
+**Output Example:**
+If analyzing features "age" and "region":
+
+| feature | group | count | error_count | error_rate | accuracy |
+|---------|-------|-------|-------------|------------|----------|
+| age | (34.667, 47.333] | 2 | 1 | 0.50 | 0.50 |
+| region | South | 2 | 1 | 0.50 | 0.50 |
+| region | North | 3 | 1 | 0.33 | 0.67 |
+| age | (21.962, 34.667] | 4 | 1 | 0.25 | 0.75 |
+| age | (47.333, 60.0] | 2 | 0 | 0.00 | 1.00 |
+| region | East | 1 | 0 | 0.00 | 1.00 |
+| region | West | 2 | 0 | 0.00 | 1.00 |
+
+*(Note: Rows with equal error_rate may appear in any order)*
+
+**Common mistakes:**
+- Passing columns in `feature_columns` that are not present in `X`.
+- Setting `min_count` too high, which may filter out all groups for some features.
+
+---
+
 ## Typical Workflow
 
 ```python
@@ -223,11 +342,12 @@ import matplotlib.pyplot as plt
 from sklearn.tree import DecisionTreeClassifier
 from ds_utils.metrics.confusion_matrix import plot_confusion_matrix
 from ds_utils.metrics.learning_curves import plot_metric_growth_per_labeled_instances
-from ds_utils.metrics.probability_analysis import visualize_accuracy_grouped_by_probability
+from ds_utils.metrics.probability_analysis import visualize_accuracy_grouped_by_probability, plot_error_analysis_chart
 from ds_utils.metrics.curves import (
     plot_roc_curve_with_thresholds_annotations,
     plot_precision_recall_curve_with_thresholds_annotations,
 )
+from ds_utils.metrics.error_analysis import generate_error_analysis_report
 
 clf = DecisionTreeClassifier(random_state=42)
 clf.fit(X_train, y_train)
@@ -267,4 +387,19 @@ fig_pr = plot_precision_recall_curve_with_thresholds_annotations(
     positive_label=1,
 )
 fig_pr.show()
+
+# 6. Error analysis chart
+y_pred = clf.predict(X_test)
+y_proba_pos = clf.predict_proba(X_test)[:, 1]
+plot_error_analysis_chart(y_test, y_pred, y_proba_pos, positive_class=1)
+plt.tight_layout()
+plt.show()
+
+# 7. Error Analysis block report
+report_df = generate_error_analysis_report(
+    X_test, y_test, y_pred,
+    feature_columns=None,
+    bins=3
+)
+print(report_df)
 ```
