+++ tests/baseline_images/test_metrics/test_probability_analysis/test_plot_error_analysis_chart_multiclass.png
