+++ tests/test_version.py
@@ -5,4 +5,4 @@ import ds_utils
 
 def test_version():
     """Test that the package version is correct."""
-    assert ds_utils.__version__ == "1.10.0rc8"
+    assert ds_utils.__version__ == "1.10.0rc9"
